requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
PyJWT==2.8.0
ruff==0.3.7
sqlparse==0.5.0
